refactor / clean workflow.go
